% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpsol-package.R
\name{glpk_version}
\alias{glpk_version}
\title{Version of the linked GLPK library}
\usage{
glpk_version()
}
\value{
A character string, e.g. `"5.0"`.
}
\description{
Version of the linked GLPK library
}
\examples{
glpk_version()
}
