% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpsol-package.R
\docType{package}
\name{glpsol-package}
\alias{glpsol}
\alias{glpsol-package}
\title{glpsol: Bundled 'GLPK' Solver for 'MathProg', 'LP' and 'MPS' Models}
\description{
Ships the 'glpsol' stand-alone solver of the GNU Linear Programming Kit ('GLPK') with R and runs it as a separate process with a time limit and an optional live log. 'glpsol' reads 'MathProg' (GMPL) models with separate data files, CPLEX LP and MPS files, and solves them with the simplex, interior-point or branch-and-cut methods, writing the solution to a file. Windows and macOS binary builds carry a static copy of 'GLPK'; other platforms link against the system library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/optimal2050/glpsol}
  \item Report bugs at \url{https://github.com/optimal2050/glpsol/issues}
}

}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

Authors:
\itemize{
  \item Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})
}

Other contributors:
\itemize{
  \item Free Software Foundation, Inc. (src/glpk/glpsol.c from 'GLPK' 5.0, see inst/COPYRIGHTS) [copyright holder]
}

}
\keyword{internal}
